refactor module names; add `serialization` subpackage
